chore: extract accelerator naming to internal/accelerator; drop dead helpers
